Invalid operation exception fix
Trying request a non-existing vendor (GET /vendors/{id}) caused the
controller to throw an invalid operation exception. This commit causes
controller to return a 404 instead. Currently a string representation of
the exception is passed to the NotFound object that is created. This is
done to help with debugging. At some point, we probably want to clean
this up.