Plumbing for Google Photos API
One endpoint (that allows user to retrieve album by album title) to test
plumbing:
GET /api/photos/albums/{id}

This is a PoC. Off the top of my head, considerations we need to account for
include (in no particular order):
- "auth"
- session "caching"
- album id (maybe something like userId+Guid/date?)
- how to correlate album to shop

Do note,
1. the following configuration options are needed
- CasCap:GooglePhotosOptions:User
- CasCap:GooglePhotosOptions:Scopes:0
- CasCap:GooglePhotosOptions:Scopes:1
- CasCap:GooglePhotosOptions:Scopes:2
- CasCap:GooglePhotosOptions:Scopes:3
- CasCap:GooglePhotosOptions:Scopes:4
- CasCap:GooglePhotosOptions:ClientId
- CasCap:GooglePhotosOptions:ClientSecret
2. because of time constraints, currently using CasCap.Apis.GooglePhotos
as a wrapper to handle OAuth2 and the nuances of interacting with Google
Photos API. The library is pretty well written. It may be fun to create
our own client library.